interval time data

+++ database/migrations/2021_01_13_013144_create_barangs_table.php
@@ -17,7 +17,7 @@ class CreateBarangsTable extends Migration
             $table->string('kode_barang')->primary();
             $table->string('barcode')->nullable();
             $table->string('nama');
-            $table->float('stok');
+            $table->double('stok', 12, 5);
             $table->timestamps();
             $table->softDeletes();
         });
